move shared resources into shared/
